fix: error handling in all endpoints, and put permission check for saving & deleting style in APIs

+++ sites/geohub/src/routes/api/style/[id]/+server.ts
@@ -1,5 +1,4 @@
 import type { RequestHandler } from './$types'
-import { error } from '@sveltejs/kit'
 import pkg from 'pg'
 const { Pool } = pkg
 
@@ -12,7 +11,9 @@ export const GET: RequestHandler = async ({ params }) => {
   try {
     const styleId = params.id
     if (!styleId) {
-      throw error(400, { message: `id parameter is required.` })
+      return new Response(JSON.stringify({ message: `id parameter is required.` }), {
+        status: 400,
+      })
     }
 
     const query = {
@@ -23,7 +24,9 @@ export const GET: RequestHandler = async ({ params }) => {
     const res = await client.query(query)
 
     if (res.rowCount === 0) {
-      throw error(404)
+      return new Response(undefined, {
+        status: 404,
+      })
     }
 
     return new Response(JSON.stringify(res.rows[0]))
@@ -37,13 +40,21 @@ export const GET: RequestHandler = async ({ params }) => {
  * Delete style.json which is stored in PostgreSQL database
  * DELETE: ./api/style/{id}
  */
-export const DELETE: RequestHandler = async ({ params }) => {
+export const DELETE: RequestHandler = async ({ params, locals }) => {
+  const session = await locals.getSession()
+  if (!session) {
+    return new Response(JSON.stringify({ message: 'Permission error' }), {
+      status: 403,
+    })
+  }
   const pool = new Pool({ connectionString })
   const client = await pool.connect()
   try {
     const styleId = params.id
     if (!styleId) {
-      throw error(400, { message: `id parameter is required.` })
+      return new Response(JSON.stringify({ message: `id parameter is required.` }), {
+        status: 400,
+      })
     }
     const query = {
       text: `DELETE FROM geohub.style WHERE id = $1`,
@@ -52,7 +63,9 @@ export const DELETE: RequestHandler = async ({ params }) => {
 
     const res = await client.query(query)
     if (res.rowCount === 0) {
-      throw error(404, { message: `${styleId} does not exist in the database` })
+      return new Response(JSON.stringify({ message: `${styleId} does not exist in the database` }), {
+        status: 404,
+      })
     }
     return new Response(undefined, {
       status: 204,
